Fix -Werror=int-to-pointer-cast/-Werror=pointer-arith across 25 files
Use (uintptr_t) intermediates for int↔pointer casts and
(uint8_t*) for void* arithmetic, making the codebase safe
on 64-bit hosts. Verified: 0 errors, all component tests pass.

Files changed:
  C++: CGWND, Game, GameObject, BuildingPanel, ScriptedObject,
       TrackPiece, TrainStation, DDRAW, LOCOBITMAP, Cursor,
       Cursor_Editor, Cursor_impls, EditWindow, GameSetupPanel,
       HelpWnd_stubs, PostcardPreviewWindow, TrainStationWindow,
       UIPANEL, UIPANEL_Surface
  C:   DDRAW_DimSurfaceRect, DDRAW_GetSurfaceWidthHeight,
       assetmgr_loadfile, cgwnd_palette, ddraw_filedata,
       multiplayer_lobby_reload

+++ src/decompiled_cpp/core/Game.cpp
@@ -308,7 +308,7 @@ Game::~Game()
     this->SetScreenMode(0, 1, 0);   /* capture=0, show=1, custom=0 */
 
     /* Release resources: vtable[6] = InitBase(0, -1, 0) = resource release call */
-    (*(void (__thiscall**)(void*, int, int, int))((int*)this)[6])(this, 0, -1, 0);
+    (*(void (__thiscall**)(void*, int, int, int))((uintptr_t*)this)[6])(this, 0, -1, 0);
 }
 
 /* ================================================================== */
